Fix clang-format misformatting
It interprets macro arguments to stringify as C
