+++ src/sources.mk
@@ -1,6 +1,6 @@
 SOURCES = mbuf.c utf.c varint.c tokenizer.c array.c boolean.c math.c string.c \
           ast.c vm.c gc.c parser.c interpreter.c slre.c object.c error.c \
-          number.c json.c main.c date.c function.c stdlib.c regex.c
+          number.c json.c main.c date.c function.c stdlib.c js_stdlib.c regex.c
 HEADERS = license.h utf.h tokenizer.h mbuf.h ast.h parser.h mm.h internal.h \
           vm.h gc.h slre.h varint.h
